+++ libraries/transformer/src/test/java/androidx/media3/transformer/SequenceExportTest.java
@@ -29,6 +29,7 @@ import static androidx.media3.transformer.TestUtil.createChannelCountChangingAud
 import static androidx.media3.transformer.TestUtil.createPitchChangingAudioProcessor;
 import static androidx.media3.transformer.TestUtil.createTransformerBuilder;
 import static androidx.media3.transformer.TestUtil.getDumpFileName;
+import static androidx.media3.transformer.TestUtil.getSequenceDumpFilePath;
 import static androidx.media3.transformer.TestUtil.removeEncodersAndDecoders;
 import static com.google.common.truth.Truth.assertThat;
 
@@ -496,6 +497,171 @@ public final class SequenceExportTest {
             /* originalFileName= */ FILE_AUDIO_RAW_VIDEO,
             /* modifications...= */ "silenceHighPitch",
             "silenceHighPitch"));
+  }
+
+  @Test
+  public void start_gapGap_completesSuccessfully() throws Exception {
+    CapturingMuxer.Factory muxerFactory = new CapturingMuxer.Factory(/* handleAudioAsPcm= */ true);
+    Transformer transformer =
+        createTransformerBuilder(muxerFactory, /* enableFallback= */ false).build();
+    EditedMediaItemSequence sequence =
+        new EditedMediaItemSequence.Builder().addGap(300_000).addGap(200_000).build();
+    Composition composition = new Composition.Builder(sequence).build();
+
+    transformer.start(composition, outputDir.newFile().getPath());
+    TransformerTestRunner.runLooper(transformer);
+
+    DumpFileAsserts.assertOutput(
+        context, muxerFactory.getCreatedMuxer(), getDumpFileName("gap", "500ms"));
+  }
+
+  @Test
+  public void start_itemGapGap_completesSuccessfully() throws Exception {
+    CapturingMuxer.Factory muxerFactory = new CapturingMuxer.Factory(/* handleAudioAsPcm= */ true);
+    Transformer transformer =
+        createTransformerBuilder(muxerFactory, /* enableFallback= */ false).build();
+    EditedMediaItem firstAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+    EditedMediaItemSequence sequence =
+        new EditedMediaItemSequence.Builder()
+            .addItem(firstAudioItem)
+            .addGap(500_000)
+            .addGap(200_000)
+            .build();
+    Composition composition = new Composition.Builder(sequence).build();
+
+    transformer.start(composition, outputDir.newFile().getPath());
+    TransformerTestRunner.runLooper(transformer);
+
+    DumpFileAsserts.assertOutput(
+        context,
+        muxerFactory.getCreatedMuxer(),
+        getSequenceDumpFilePath(
+            ImmutableList.of(FILE_AUDIO_RAW.replace("wav/", ""), "gap_500ms", "gap_200ms")));
+  }
+
+  @Test
+  public void start_gapGapItem_completesSuccessfully() throws Exception {
+    CapturingMuxer.Factory muxerFactory = new CapturingMuxer.Factory(/* handleAudioAsPcm= */ true);
+    Transformer transformer =
+        createTransformerBuilder(muxerFactory, /* enableFallback= */ false).build();
+    EditedMediaItem audioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+    EditedMediaItemSequence sequence =
+        new EditedMediaItemSequence.Builder()
+            .addGap(200_000)
+            .addGap(500_000)
+            .addItem(audioItem)
+            .build();
+    Composition composition = new Composition.Builder(sequence).build();
+
+    transformer.start(composition, outputDir.newFile().getPath());
+    TransformerTestRunner.runLooper(transformer);
+
+    DumpFileAsserts.assertOutput(
+        context,
+        muxerFactory.getCreatedMuxer(),
+        getSequenceDumpFilePath(
+            ImmutableList.of("gap_200ms", "gap_500ms", FILE_AUDIO_RAW.replace("wav/", ""))));
+  }
+
+  @Test
+  public void start_itemGapGapItem_completesSuccessfully() throws Exception {
+    CapturingMuxer.Factory muxerFactory = new CapturingMuxer.Factory(/* handleAudioAsPcm= */ true);
+    Transformer transformer =
+        createTransformerBuilder(muxerFactory, /* enableFallback= */ false).build();
+    EditedMediaItem firstAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+    EditedMediaItem secondAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+
+    EditedMediaItemSequence sequence =
+        new EditedMediaItemSequence.Builder()
+            .addItem(firstAudioItem)
+            .addGap(500_000)
+            .addGap(200_000)
+            .addItem(secondAudioItem)
+            .build();
+    Composition composition = new Composition.Builder(sequence).build();
+
+    transformer.start(composition, outputDir.newFile().getPath());
+    TransformerTestRunner.runLooper(transformer);
+
+    DumpFileAsserts.assertOutput(
+        context,
+        muxerFactory.getCreatedMuxer(),
+        getSequenceDumpFilePath(
+            ImmutableList.of(
+                FILE_AUDIO_RAW.replace("wav/", ""),
+                "gap_500ms",
+                "gap_200ms",
+                FILE_AUDIO_RAW.replace("wav/", ""))));
+  }
+
+  @Test
+  public void start_itemGapItemGap_completesSuccessfully() throws Exception {
+    CapturingMuxer.Factory muxerFactory = new CapturingMuxer.Factory(/* handleAudioAsPcm= */ true);
+    Transformer transformer =
+        createTransformerBuilder(muxerFactory, /* enableFallback= */ false).build();
+    EditedMediaItem firstAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+    EditedMediaItem secondAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+
+    EditedMediaItemSequence sequence =
+        new EditedMediaItemSequence.Builder()
+            .addItem(firstAudioItem)
+            .addGap(200_000)
+            .addItem(secondAudioItem)
+            .addGap(500_000)
+            .build();
+    Composition composition = new Composition.Builder(sequence).build();
+
+    transformer.start(composition, outputDir.newFile().getPath());
+    TransformerTestRunner.runLooper(transformer);
+
+    DumpFileAsserts.assertOutput(
+        context,
+        muxerFactory.getCreatedMuxer(),
+        getSequenceDumpFilePath(
+            ImmutableList.of(
+                FILE_AUDIO_RAW.replace("wav/", ""),
+                "gap_200ms",
+                FILE_AUDIO_RAW.replace("wav/", ""),
+                "gap_500ms")));
+  }
+
+  @Test
+  public void start_gapItemGapItem_completesSuccessfully() throws Exception {
+    CapturingMuxer.Factory muxerFactory = new CapturingMuxer.Factory(/* handleAudioAsPcm= */ true);
+    Transformer transformer =
+        createTransformerBuilder(muxerFactory, /* enableFallback= */ false).build();
+    EditedMediaItem firstAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+    EditedMediaItem secondAudioItem =
+        new EditedMediaItem.Builder(MediaItem.fromUri(ASSET_URI_PREFIX + FILE_AUDIO_RAW)).build();
+
+    EditedMediaItemSequence sequence =
+        new EditedMediaItemSequence.Builder()
+            .addGap(500_000)
+            .addItem(firstAudioItem)
+            .addGap(200_000)
+            .addItem(secondAudioItem)
+            .build();
+    Composition composition = new Composition.Builder(sequence).build();
+
+    transformer.start(composition, outputDir.newFile().getPath());
+    TransformerTestRunner.runLooper(transformer);
+
+    DumpFileAsserts.assertOutput(
+        context,
+        muxerFactory.getCreatedMuxer(),
+        getSequenceDumpFilePath(
+            ImmutableList.of(
+                "gap_500ms",
+                FILE_AUDIO_RAW.replace("wav/", ""),
+                "gap_200ms",
+                FILE_AUDIO_RAW.replace("wav/", ""))));
   }
 
   @Test